# api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/test":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["test"]}},"/api/tasks":{"post":{"operationId":"TasksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskEntity"}}}}},"tags":["tasks"],"security":[{"bearer":[]}]},"get":{"operationId":"TasksController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskEntity"}}}}}},"tags":["tasks"],"security":[{"bearer":[]}]}},"/api/tasks/{id}":{"get":{"operationId":"TasksController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskEntity"}}}}},"tags":["tasks"],"security":[{"bearer":[]}]},"patch":{"operationId":"TasksController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskEntity"}}}}},"tags":["tasks"],"security":[{"bearer":[]}]},"delete":{"operationId":"TasksController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskEntity"}}}}},"tags":["tasks"],"security":[{"bearer":[]}]}},"/api/auth/sign-up":{"post":{"operationId":"AuthenticationController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["auth"]}},"/api/auth/sign-in":{"post":{"operationId":"AuthenticationController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninResponse"}}}}},"tags":["auth"]}},"/api/auth/refresh-tokens":{"post":{"operationId":"AuthenticationController_refeshTokens","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninResponse"}}}}},"tags":["auth"]}},"/api/auth/google":{"post":{"operationId":"GoogleAuthenticationController_authenticate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoogleTokenDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninResponse"}}}}},"tags":["auth/google"]}}},"info":{"title":"Taskify API","description":"The Taskify API description","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateTaskDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"assignedToId":{"type":"number"},"startDate":{"type":"string"},"endDate":{"type":"string"}},"required":["title","status","assignedToId","startDate","endDate"]},"TaskEntity":{"type":"object","properties":{"id":{"type":"number"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"title":{"type":"string"},"description":{"type":"string"}},"required":["id","startDate","endDate","title","description"]},"UpdateTaskDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"assignedToId":{"type":"number"},"startDate":{"type":"string"},"endDate":{"type":"string"}}},"SignUpDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"phone":{"type":"string"}},"required":["firstName","lastName","password"]},"UserEntity":{"type":"object","properties":{"id":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"googleId":{"type":"string"}},"required":["id","firstName","lastName","phone","email","role","googleId"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"SigninResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"GoogleTokenDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]}}}}